app/requirements.txt
fastapi==0.115.4
uvicorn==0.32.0
pillow==10.4.0
sqlalchemy==2.0.36
alembic==1.13.3
pydantic==2.9.2
asyncpg==0.30.0
psycopg2-binary==2.9.10
celery==5.4.0
redis==5.2.0
python-dotenv==1.0.1
python-jose==3.3.0
passlib[bcrypt]==1.7.4
pika==1.3.1
